feat: standardize package naming and add config
- Update package name to 'laravel-logpond' for consistency
- Introduce default package configuration file
- Add 'logging' keyword to composer.json for better discoverability

Renaming the package standardizes its identity across the repository, making it more intuitive for developers to recognize and utilize. The addition of a configuration file provides users with the ability to customize the package without modifying the core code, enhancing flexibility and ease of use. The updated keywords in composer.json improve the package's visibility and relevance in searches related to Laravel logging solutions.